requirements.txt
#Install Command
#python -m pip install -r requirements.txt

# Flask and Extensions
Flask==2.0.1
Flask-Session==0.4.0
Werkzeug==2.0.1

# HTTP and API
requests==2.26.0
requests-oauthlib==1.3.0
urllib3==1.26.7

# Environment and Config
python-dotenv==0.19.0
PyYAML==5.4.1

# Security
itsdangerous==2.0.1
cryptography==3.4.8

# Development
black==21.9b0
flake8==3.9.2

# Production
gunicorn==20.1.0

# Utilities
python-dateutil==2.8.2
pytz==2021.3

# Testing Dependencies
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0